SCrypt: Extracted `Engine`, `Errors` and `Password` into dedicated files...
- Engine: Replaced `scanf` usage to avoid adding a runtime dependency for upcoming Ruby 2.7, replaced various nested `if/else` blocks in favor of guard-clause to improve code readability, moved all `self.` methods into `self` class.

- Errors + Passwords: Relocated into their own files and formatting (no logic change)
